Update usePostLogin location
