+++ shared/common/src/components/Auth/Login/LoginForm/LoginButtons/index.tsx
@@ -1,60 +1,23 @@
 'use client'
 
-import { TokenManager, usePostLogin } from '@bitgouel/api'
+import { TokenManager } from '@bitgouel/api'
 import {
-  EmailErrorText,
   EmailValue,
-  PasswordErrorText,
-  PasswordValue,
+  LoadingStateContext,
+  PasswordValue
 } from '@bitgouel/common'
-import { useRouter } from 'next/navigation'
-import { useRecoilValue, useSetRecoilState } from 'recoil'
+import { useContext } from 'react'
+import { useRecoilValue } from 'recoil'
 import * as S from './style'
 
-const LoginButtons = () => {
-  const tokenManager = new TokenManager()
-  const { push } = useRouter()
+interface Props {
+  mutate: () => void
+}
 
+const LoginButtons = ({ mutate }: Props) => {
+  const isLoading = useContext(LoadingStateContext)
   const emailValue = useRecoilValue(EmailValue)
   const passwordValue = useRecoilValue(PasswordValue)
-  const setEmailErrorText = useSetRecoilState(EmailErrorText)
-  const setPasswordErrorText = useSetRecoilState(PasswordErrorText)
-
-  const { mutate, isLoading } = usePostLogin({
-    onSuccess: (data) => {
-      tokenManager.setTokens(data)
-      push(`/`)
-    },
-    onError: ({ response }) => {
-      if (response?.status === 404) {
-        setEmailErrorText('등록되지 않은 계정입니다.')
-      } else if (response?.status === 403) {
-        setEmailErrorText('아직 회원가입 대기중인 계정입니다')
-      } else if (response?.status === 401) {
-        setPasswordErrorText('비밀번호가 일치하지 않습니다')
-      } else if (response?.status === 400) {
-        if (
-          Object.keys(response?.data.fieldError).includes('email') &&
-          Object.keys(response?.data.fieldError).includes('password')
-        ) {
-          setEmailErrorText('잘못된 이메일입니다')
-          setPasswordErrorText('잘못된 비밀번호입니다')
-        } else if (
-          Object.keys(response?.data.fieldError).includes('email') &&
-          !Object.keys(response.data.fieldError).includes('password')
-        ) {
-          setEmailErrorText('잘못된 이메일입니다.')
-          setPasswordErrorText('')
-        } else if (
-          Object.keys(response?.data.fieldError).includes('password') &&
-          !Object.keys(response?.data.fieldError).includes('email')
-        ) {
-          setPasswordErrorText('잘못된 비밀번호입니다.')
-          setEmailErrorText('')
-        }
-      }
-    },
-  })
 
   const disabled = (): boolean =>
     isLoading || !emailValue.length || !passwordValue.length
@@ -66,7 +29,7 @@ const LoginButtons = () => {
       <S.LoginButton
         disabled={disabled()}
         isAble={isAble()}
-        onClick={() => mutate({ email: emailValue, password: passwordValue })}
+        onClick={mutate}
       >
         로그인
       </S.LoginButton>
